+++ reverse_proxy/Cargo.toml
@@ -18,3 +18,4 @@ tokio-native-tls = { workspace = true }
 tokio-util = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+response = { path = "../response" }
